backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
umap-learn
openai
GitPython
networkx
scikit-learn
numpy
python-dotenv
